Changed format of challenges
I wanted to change the format of grabbing challenges and displaying them to users on the server. There wasn't much information being stored in the database, so I changed it so that the challenges are separated by category and including information such as the name, value, whether it was solved or not, and who it was solved by (blank if unsolved). This way, I can store point values and solver information and display it for the team to see in the discord server. Also removed the gen_page method because I did not want to use it with the new implementation.