Removed unnecessary stuff and made pull static
Removed methods we did not want to use for the bot (add/remove/status) and functions associated to them. Removed the pull command and instead changed it to be a helper method of list. Every time list is called now, it will grab the information from scratch, thus updating the challenge info. This means there is no need for people to change status of challenges as it can be pulled from the DB.