Cleaned and edited functions
Created creds dictionary locally in script to keep track of credentials. This makes it so that I don't have to store the credentials in the database and so every user in the server can see the user's password. I made it so that users have to DM the bot in order to set the credentials. I also removed the getcreds and creds functions because I don't think it should be used by members. There might be more functionality down the lines for single-account users, but for now, this is it. Changed the name of the list command top challenges to make it shorter and to not have extra groups of commands for no reason. Edited the help_info.py to accommodate my changes. Also added a rating dictionary for each member of a server. I also made it so that each server the bot is on has it's own database. I wanted to make a database inside of a database, but I don't think it's even possible.