Fixed setcreds
I fixed the setcreds command so that only members can DM the bot in order to set creds. This is used so that no other members can see the password of the credentials, BUT they can see that credentials were set so that they don't have to do it themselves.