rCTF Integration
Found a way to get the API info for the rCTF platform. It relies on a token stored in the Local Storage of the browser. There isn't a very solid way of getting this token, but it seems the one I put in the script manually works for the time being. Will probably need a flexible workaround to it eventually.