Editing scoring system; update >ctf archive
The scoring system is actually going to work now as I found a solution. Will be represented in the README.md at some point. The archive method is going to start acting like the deep_archive method so that we tarball each channel individually. It will be easier to search through messages this way and we don't have to worry about manually running >deep_archive which can get to be annoying